Community Time Series Comparison

Compute properties of community time series generated with generateTS.

compareTS(input.folder = "", expIds = c(), modif.folder = "",
  modif = "", testNeutral = FALSE, sliceDef = c(1, NA), epsilon = 0.2,
  predef = FALSE, detrend = TRUE, norm = FALSE, hurstBins = c(0.5, 0.7,
  0.9), maxautocorBins = c(0.3, 0.5, 0.8), timeDecaySliceDef = c(1, 50),
  varEvolSliceDef = c(), radSliceDef = c(), returnDistribs = FALSE,
  returnTS = FALSE)

Arguments

input.folder
the folder where results (settings and time series) of function generateTS are stored.

expIds
the experiment identifiers of time series to be considered

modif.folder
the folder with time series sub-sets as generated by sliceTS.R or with noisy time series as generated by addNoise.R (settings are read in from input.folder)

modif
modification (shared name of time series in modif.folder, sliced for time series sub-sets and pois for time series with Poisson noise added)

testNeutral
test neutrality (requires libraries WrightFisher and logitnorm)

sliceDef
consider the time series sub-set defined by the given start and end time point for property computation, an end time point of NA means that the entire time series is considered; timeDecaySliceDef and varEvolSliceDef refer to the outcome of sliceDef

epsilon
allowed deviation from the expected slope for noise type identification

predef
recognize noise types with predefined slope boundaries (see identifyNoisetypes for details)

detrend
remove linear trends before computing the periodogram (recommended)

norm
normalize time series by dividing each entry in a sample by the sum of its sample; if norm is true, entropy computation is omitted

hurstBins
binning thresholds for Hurst exponent (three thresholds required for four bins)

maxautocorBins
binning thresholds for maximal autocorrelation (three thresholds required for four bins)

timeDecaySliceDef
the time series subset considered to compute the time decay, a vector with the start and the end time point (if the end time point is NA, the entire time series is used)

varEvolSliceDef
(optional, by default not computed) the time series subset considered to compute the evolution of variance, a vector with the start and the end time point (if the end time point is NA, the entire time series is used)

radSliceDef
(optional, by default not computed) the time series subset considered to analyse the (non-normalized also when norm is enabled) species abundance versus rank curve, a vector with the start and the end time point (if the end time point is NA, the entire time series is used); cannot be used together with sliceDef

returnDistribs
return distributions at the final time point (no time series properties are computed, cannot be used together with returnTS)

returnTS
return the time series (no time series properties are computed, cannot be used together with returnDistribs)

Value

a table with experiment parameters (algorithm, connectance, sigma, theta and so on) and time series properties (noise types, slope of Taylor's law etc.); if returnDistrib is true, a list with the abundances at the last time point, if returnTS is true, a list with the time series

Details

If infotheo is installed, the entropy will be also computed. Noise types are computed with smooth set to true. The neutrality test is carried out with ntests set to 500 and method logitnorm.