Dissimilarity-Overlap Curve (DOC)

Given matrix of relative taxon abundances, plot the dissimilarity-overlap curve (DOC) sample-wise. The DOC method was developed by Bashan and colleagues.

doc(x, B = 100, polygons = FALSE, rand = FALSE, lower.conf = 0.03,
  upper.conf = 0.97, null.model = "assembly", doc.res = NULL)

Arguments

x
rows are taxa, columns are samples, abundance in each sample is assumed to sum to one

B
bootstrap iterations, if set to 0 or below, no bootstraps are carried out

polygons
draw a polygon instead of the lower and upper confidence line (in development)

rand
randomize the data sample-wise before starting the DOC analysis

lower.conf
lower limit of the confidence interval

upper.conf
upper limit of the confidence interval

null.model
the null model to use, permut shuffles x sample-wise, assembly selects for each non-zero taxon one of the values taken across the samples at random

doc.res
the result of a previous run of the doc method is added in green with black points, current result is shown in red with grey points (to plot null model together with data)

Value

a list with the overlaps, dissimilarities, lowess smoothed overlaps and dissimilarites and lower and upper confidence intervals

References

A. Bashan et al. (2016). Universality of human microbial dynamics, Nature 534, 259-262.

Examples

if (FALSE) {
data("david_stoolA_otus")
data=normalize(rarefyFilter(david_stoolA_otus,min=10000)[[1]])
doc.out=doc(data[,1:50],B=10) # apply the DOC method to the first 50 samples
# plot the DOC curve together with the curve for the null model
doc.null=doc(data[,1:50],B=10,rand=TRUE,doc.res=doc.out)
}