LIMITS

LIMITS is an algorithm developed by Fisher & Mehta to estimate the interaction matrix from time series data assuming a Ricker model.

limits(x, bagging.iter = 100, verbose = FALSE)

Arguments

x
time series with taxa as rows and time points as columns

bagging.iter
the number of iterations used for bagging

verbose
print which taxon LIMITS is processing

Value

the estimated interaction matrix

References

Fisher & Mehta (2014). Identifying Keystone Species in the Human Gut Microbiome from Metagenomic Timeseries using Sparse Linear Regression. PLoS One 9:e102451

Examples

not_run({
  N=20
  A=generateA(N,c=0.1)
  ts=ricker(N=N,A=A)
  Aest=limits(ts,verbose=TRUE)
  par(mfrow=c(2,1))
  plotA(A,header="original")
  plotA(Aest,header="estimated")
  par(mfrow=c(1,1))
})