Compute evenness using Sheldon's index

Sheldon's index is defined as \(S=\frac{e^H}{N}\), where H is the Shannon diversity and N the species number. It ranges from 0 to 1, where 1 signifies a perfectly even abundance distribution.

sheldon(x, correction = TRUE, N2N1 = FALSE)

Arguments

x
a vector of species abundances

correction
whether or not to apply the correction described in Alatalo, Oikos 37, 199-204, 1981

N2N1
whether to compute Sheldon's evenness as the ratio of e raised to the power of H (H = Shannon diversity) and Simpson's diversity

Value

Sheldon's evenness

Details

Note that the N2N1 mode results in evenness smaller than 1 for equal taxon probabilities.

References

A.L. Sheldon 1969. Equitability indices: dependence on the species count. Ecology, 50, 466-467.

C Heip 1974. A new index measuring evenness. J. mar. biol. Ass. UK 54, 555-557.

Examples

N=50
# uneven species abundance distribution
v = round(generateAbundances(N, mode=5))
sheldon(v)

#> [1] 0.6743761

# perfectly even species abundance distribution
sheldon(rep(N,N))

#> [1] 1

# very uneven species abundance distribution
sheldon(c(rep(1,N),1000))

#> [1] 0.009179405