Slice time series

Store slices of previously generated time series in the input folder and store them in the output folder.

sliceTS(slices = NULL, slice.def.path = "", input.folder = "",
  output.folder = "", expIds = c())

Arguments

slices
a matrix with rows and 2 columns, where each row codes the time series identifier, the first column the start and the second the end time point; NA means until end of the time series

slice.def.path
path to a slice definition file, which has two tab-delimited columns for the start and end time points, respectively; NA means until end of the time series

input.folder
location of time series and settings sub folders

output.folder
folder in which all slices go (no sub-folders)

expIds
set of experiment identifiers to process

Details

The input folder is expected to contain the folder structure as generated by generateTS.R (timeseries subfolder containing experiment subfolders). Sliced time series are stored in the output folder without subfolders.