Generate community time series

Wrapper function for time series generation. The result folder has a specific structure that is created if it does not yet exist. It has two sub-folders, one named timeseries and the other settings. In both sub-folders, there are folders having as name the experiment identifier followed by timeseries or settings. The interaction matrix, immigration and extinction rates, carrying capacities and other randomly generated input are also stored in settings. The idea is that the time series can be re-created from the information stored in the settings folder (though not exactly, since the SOI, Ricker and neutral models have random components). Furthermore, a plot of the time series and (if required by the model) the interaction matrix is stored in settings, for control. Properties of the interaction matrix are also stored there.

generateTS(N = 100, I = 1500, tend = 100, initAbundMode = 5, c = 0.05,
  clique.size = 5, sigma = 0.05, theta = 0.002, Atype = "klemm",
  Atweak = "tweak", d = -1, PEP = 30, m = 0.02, deathrate = 10,
  algorithm = "dm", interval = 1, output.folder = "", output.expId = "",
  input.folder = "", input.expId = "", read.A = FALSE, read.y = FALSE,
  read.K = FALSE, read.extinct = FALSE, read.ts = FALSE, maxIter = 10)

Arguments

N
species number (for hubbell, refers to the species number in the metacommunity)

I
individual number (only for SOI and hubbell)

tend
the number of time points to generate

initAbundMode
the mode in which initial abundances are generated (described in function generateAbundances)

c
connectivity of interaction matrix (only for SOI, ricker and gLV)

clique.size
parameter for interaction matrix generation (only for SOI, ricker and glv and Atype mode klemm)

sigma
noise term (only for ricker)

theta
Dirichlet-Multinomial overdispersion parameter (only for dm)

Atype
method to generate the interaction matrix (described in function generateA)

Atweak
tweaking method for instable interaction matrix, either tweak (convert positive into negative arcs until the matrix is stable or one third of the positive arcs have been converted) or schur (apply Schur decomposition, does not guarantee a stable matrix)

d
diagonal value of the interaction matrix

PEP
positive edge percentage of interaction matrix (only for SOI, ricker and gLV)

m
immigration rate (only for hubbell)

deathrate
number of individuals that are replaced at each time step (only for hubbell)

algorithm
how time series should be generated (can be hubbell, soi, ricker, glv, dm, davida or davidb)

interval
sampling frequency (if 1, each sample is taken, if 2, every second sample is taken etc.)

output.folder
path to result folder, if specified, input parameter and time series are exported to the output folder

output.expId
the identifier of the time series generation experiment

input.folder
the folder from which previously generated parameters can be read

input.expId
the identifier of a previous experiment whose results will be read

read.A
read the interaction matrix from the given previous experiment

read.y
read the initial abundances/SOI immigration rates from the given previous experiment (for hubbell, initial abundances refer to the metacommunity)

read.K
read the carrying capacities/growth rates from the given previous experiment

read.extinct
read the SOI extinction rates from the given previous experiment

read.ts
read the previously generated time series from the given previous experiment

maxIter
maximal iteration number for finding a stable interaction matrix

Value

list containing the generated interaction matrix, initial abundances (they double as immigration probabilities for soi and metapopulation proportions for hubbell), carrying capacities (they double as growth rates for glv), extinction probabilities, time series and settings

Details

On the random generation of parameter values: Extinction rates are sampled from the uniform distribution with values between 0 and 1. Carrying capacities/growth rates are sampled from the uniform distribution with values between 0 and 0.5. On the implementation of the neutral model: simHubbell is used to generate neutral model time series. The species number of the local community is set to 1/10th of the species number in the metacommunity. The local community is initialized with even abundances.