Get the taxonomy given OTU names and lineage information

The lineage information is provided in form of a matrix, which contains for each OTU identifier the taxonomic levels. OTUS in the lineage matrix are rows, taxonomic levels columns.

getTaxonomy(selected = c(), lineages, level = "class")

Arguments

selected
a character vector of selected OTU identifiers

lineages
a lineage table

level
the taxonomic level (domain, phylum, class, order, family, genus or species)

Value

the taxonomy of the OTUs

Examples

data(david_stoolA_otus)
data(david_stool_lineages)
sorted=sort(apply(david_stoolA_otus,1,sum),decreasing=TRUE)[1:10]
getTaxonomy(names(sorted),david_stool_lineages,level="family")

#>  [1] Ruminococcaceae     Bacteroidaceae      Erysipelotrichaceae
#>  [4] Bacteroidaceae      Ruminococcaceae     none               
#>  [7] Lachnospiraceae     Bifidobacteriaceae  Lachnospiraceae    
#> [10] Lachnospiraceae    
#> 152 Levels: A29 ACK-M1 AKIW874 Acetobacteraceae ... none