Hubbell Simulation

Simulate species abundances with the basic Hubbell model.

simHubbell(N = 50, M = 500, I = 500, y = rep(1/N, N),
  m.vector = rep(1/M, M), m = 0.02, d = 10, tskip = 0, tend = 100)

Arguments

N
species number in the local community

M
species number in the metacommunity

I
number of individuals

y
initial species probabilities, should sum to one

m.vector
species proportions in the metacommunity, should sum to one

m
immigration rate (probability that a dead individuum is replaced by an individuum from the metacommunity)

d
number of deaths at each time step

tskip
number of initial time points to be skipped when returning the result (to avoid the transient)

tend
number of time points (i.e. the number of generations)

Value

a matrix with species abundances as rows and time points as columns

Details

For a recent review on the Hubbell model, see Rosindell, Hubbell and Etienne, 2011.

References

Rosindell, Hubbell and Etienne (2011). The Unified Neutral Theory of Biodiversity and Biogeography at Age Ten. Trends in Ecology and Evolution, vol. 26 (7), 340-348.

See also

simUntb for the neutral model with the untb package

Examples

not_run({
    N <- 50
    M <- 500
    metapop <- generateAbundances(N = M, mode = 5, probabs = TRUE)
    x <- tsplot(simHubbell(N = N, M = M, I = 3000, d = N, m.vector = metapop, tskip=500, tend=1000))
})