Compute statistics on jumps through community space

Consecutive samples representing time points in an ordination plot can be interpreted as vectors, which have a length and an angle. If ordinate is true, these lengths and angles are computed. If ordinate is false, dissimilarities between consecutive time points are computed. If a perturbation object is provided and plot.type hist is true, the histogram of jump lengths is plotted and the significance of the difference between perturbed and non-perturbed jump lengths is assessed with a Wilcoxon test.

tsJumpStats(x, time.given = FALSE, time.unit = "days", ordinate = TRUE,
  distance = "bray", dimensions = c(1, 2), plot = FALSE,
  plot.type = "jumps", header = "", subsample = FALSE, perturb = NULL)

Arguments

x
a community time series matrix, with rows as taxa and columns as time points

time.given
sample names provide time steps, only needed for plotting and plot.type jumps

time.unit
unit of time, only needed for plotting and plot.type jumps

ordinate
if TRUE, compute jumps in ordination plot

distance
the distance (or dissimilarity) used to compute jumps directly or to compute the ordination

dimensions
if ordinate is TRUE, the principal components considered

plot
if TRUE, a plot of the jumps is made

plot.type
jumps: plot jumps as bars in the order of occurrence, hist: a histogram of jump lengths for perturbed and non-perturbed sample pairs (only if perturbation object is provided)

header
text to be used as plot title instead of default text

subsample
for plot.type hist: subsample larger jump vector randomly down to smaller one, so that there are as many perturbed as non-perturbed jumps

perturb
a perturbation object, if provided, the plot is colored accordingly

Examples

not_run({
  data(david_stoolA_otus)
  data(david_stoolA_metadata)
  rarefaction=rarefyFilter(david_stoolA_otus,min = 10000)
  rarefiedA=rarefaction$rar
  days=david_stoolA_metadata[1,rarefaction$colindices]
  interpA=interpolate(rarefiedA,time.vector=days,interval=1,method="stineman")
  interpA[interpA<0]=0
  perturbA=perturbation(times=c(80,105), durations=c(5,9))
  res=tsJumpStats(interpA, plot=TRUE, perturb=perturbA, header="in stool A")
  out=powerspec(res$lengths,plot=TRUE)
})